added functionality for more specific jam packets
jam packets now have 2 optional arguments. The first is an integer being the type of sensor to target. If it is not included or 0, then all sensors types are jammed. The second is an integer being the id of the sensor to target. If it is not included or 0, then all sensor ids are jammed.